Bug fixes
ExpertConnectDemo - Logout was not completely logging out the user. Now
calls [userManager unauthenticateUser].

HorizonDemo - plist file added to project workspace. Added
ExceptionDomains for building in Xcode7.

SDK - Corrected case where server request returned a dictionary
(DecisionEngine) that was not a known actionType. Now passes it along
as a NSDictionary instead of crashing.

SDK - Chat startup will now check for a missing userToken and use the
deviceID if it is missing. This corrects the issue where chat when
using “Skip Registration” in the demo app would not start.

CafeX - Now bails on attempting to connect if “cafeXHost” configuration
parameter is empty.