Organization Selector
Added an organization selector to the settings view in the demo app.
The choices are dependent on which environment is selected. The user’s
choice for each environment is saved.

The SDK now has a “setClientID” function which allows the integrator to
select a different clientID. This will blow away the authToken so that
the next call to the API will negotiate a new token using the new
clientID. There is no current use case where an integrator would need
this but it is there so the demo app can switch clients.