Corrected crash when receiving HTML or XML response
Some errors come back from the server as HTML or XML and won’t have a
JSON structure to them. The trouble debug line was added a few versions
ago to troubleshoot an unrelated crash. That crash can occur if an
organization is sent to the server that does not exist in the
environment.