UX Config & QA Bug Fixes
Chat “send” button will be visibly disabled (50% transparent) when you
cannot sent a chat or there is no text.
Breadcrumbs will be dispatched to server upon starting a chat. This
makes sure the agent sees up-to-date breadcrumbs when the chat session
is launched.
Added function override for startAnswerEngine which provides the
“showSearchBar” parameter.
Cleaned up descriptions on ECDemo breadcrumbs.
Corrected crash if demo JSON was changed to a smaller number of forms
available.