Test Harness - Bug - Wrong skill when starting a chat
Corrected issue where starting a chat on the “Test Chat” view would
incorrectly use the first skill instead of the one that appears to be
selected. It should not correctly use the displayed skill in all cases.
(A workaround was to move the slider up & back down, this is no longer
necessary).