PAAS-1603,1605 Added "are you sure" dialog if user tries to leave chat queue
Chat - a new dialog has been added to ask the user if they are sure
they want to leave the chat queue. This dialog is displayed while the
user is in the queue for a chat. The dialog has four new localized
strings that can be overridden. The keys and default values are:
"ECSLocalizedLeaveQueueTitle" = "Leave Chat Queue?"
"ECSLocalizedLeaveQueueMessage" = "By leaving now, you will lose your
place in the chat queue."
"ECSLocalizedLeaveQueueYes" = "Yes, Leave"
"ECSLocalizedLeaveQueueNo" = "No, stay"

Chat - A bug was fixed that would cause the chat client to see a
disconnection message if a "notify" channel state was sent from the
server. This "notify" message is not currently sent by Humanify API and
is reserved for future capability.