PAAS-1603 Fix to new in-queue dialog
SDK:
Corrected the in-queue “are you sure” dialog from being displayed while
in chat and and the agent left the chat.
ETA wait time logic was modified to be exactly the same as Android.
There should be no changes due to this.

Test Harness:
Updated the ETA messaging to read more correctly.