PAAS-1740 App Crash if server not reachable
Chat will now end with a dialog if server is down and user is in queue.
Corrected crashes in multiple places in the code. These occur if the
server is not reachable (jetty down for example). Most of the issues
had to do with code checking if the error object was nil which is not
always sufficient to proceed using object subfields. You also need to
check that the incoming object is of the expected type. In this
situation, the return data was an HTML 503: Gateway error that was not
in JSON and thus not translated into a data object.