PAAS-1988 - Form previous "option" value blanked issue
Forms - Corrected issue that would cause a user's answer to not be
displayed if they clicked "previous" and the previous question was an
option list. Now the SDK should correctly maintain the user's answer
and display it properly.
Forms - Corrected issue with the "slider" form type that would cause
the "next" button to be disabled if the user selected the minimum or
maximum value.