PAAS-2223 Corrected chat disconnecting due to heartbeat issue
The server only sends a heartbeat after 20 seconds of no traffic (if
the heartbeat setting is 20000, which the SDK always sets it to.
If the user or agent constantly chat, never letting 20 seconds of idle
occur, no heartbeat will ever be sent from the server.
Currently, the iOS SDK is only checking for heartbeats to determine if
server is alive.
The fix is that ANY chat message should reset the heartbeatMiss counter.